Restructed canopy bit
